Change the lifetime parameter from `'a` to `'h`, per nmatsakis's recommendation.
Actually, Niko recommended `'heap`. Then, to my objection that you have to type
it so often that it really should be short, he suggested using `'heap` for
types and `'h` for functions. But it's bad to use two different names for
exactly the same (rather tricky) concept, so `'h` it is.