Run destructors on heap objects in-place.
Previously we copied every value out of the heap (using `ptr::read()`)
in order to drop it. This was pretty silly. It seems "obvious" that
it's just as safe to drop the values in-place; if destructors can chase
heap-to-heap edges, we already have a bad bug. I think I just didn't
know about `ptr::drop_in_place()` when I originally wrote this.

Similarly, drop page headers without copying.