Change PageSet::pages from a Vec<PageBox> to a linked list.
The point of this is, later on, to have several lists.

For example, currently, each allocation walks the list of pages until it
finds one with a nonempty freelist. As a result, after the first page in
the list fills up, every allocation is slow. That overhead can be
eliminated if we keep full and nonfull pages in separate lists.

Similarly, we won't have to sweep empty pages if we keep empty pages in
a third separate list.