impl Send for Heap.
This design decision is a real fork in the road. It means everything you
put into a `Heap` must also be `Send`. `Rc`, for example, has to be
banned. I changed the relevant tests to use `Arc`. (It would be possible
to add a `gc_cell::Rc<'h, T>` type, where `T: Send + 'static`. Maybe
later.)