Implement proper tail calls through trampolining
Most all the evaluation related function return a `Trampoline` instead of a
`Value` now. A `Trampoline` is either a completely evaluated `Value`, or a the
tail call continuation of a partial evaluation. To complete a `Trampoline`'s
evaluation, we take care to unwind and return to a non-recursive while loop that
drives the `Trampoline` to completion without blowing the stack.