Remove the lifetime parameter from Storage types.
The hack needed in gc-cell-derive to make this work is moderately gross,
but Storage types really should not have a lifetime parmater—it makes no
sense.

The IntoHeapBase::In associated type now has a 'static bound.

Support for PhantomData is reined in. I hope to delete this entirely,
eventually; the only reason to use a PhantomData field in the heap is
issue #21.