lisp: New hand-coded high-level Scheme parser.
This is not faster. It uses a lot less system stack, though, making
profiles much easier to read. Currently, parsing psyntax:
* 60% of parse time is in the Nom lexer for tokens.
* 17% of parse time is in GcHeap allocation.
* 10% of parse time is in unpinning.
*  4% of parse time is in InternedString::get().
* That leaves 9% which is presumably the overhead of my new parser;
    some of that is in creating Strings that aren't really necessary.
    But the parser is also just not super efficient.

Regarding unpinning: The first-order problem is that unpinning involves
computing a SipHash (lol). But e.g. `pair.set_cdr(value)` requires
checking if value contains a `GcRef` (pinned pointer) in which case it
needs to be unpinned because we're moving it into the heap. Ugh.