Rewrite pinning to be much faster.
As a bonus, we stop using bit_vec for mark bits and in-use bits.

This commit reduces the total run time of nboyer.scm, at level 0,
from 8.1 seconds to 2.9 seconds, a 64% improvement.

The cost is memory overhead: a word per object for a pin count. (Scheme
is just about worst-case for this. Most allocations are pairs.)