lisp: Reorganization.
The aim here was to get the three stages of the pipeline—`parse`,
`compile`, and `vm`—to be as independent as possible.
`compile` does not depend on `parse` but instead looks at `Value`s.
`vm` does not call any `compile` functions but instead looks at `Expr`s.

*   The `toplevel` module contains code that puts together parts from
    several stages. That includes the `repl` as well as `eval` and
    `eval_str`.

*   `Environment` and `StaticEnvironment` get their own module,
    `env`, and more convenient constructors.

Everything still kind of depends on everything else—this was a mixed
success.