Implement a new compiler and interpreter.
This one is bytecode-like, and it's something like 25% faster. I think
it'll also make continuations (on the one hand) and further speed
improvements (on the other) easier to implement.

The cost of constant into_heap and from_heap calls is, understandably,
terrible. Right now, 6.8% of nboyer is spent in from_heap methods, and
6.9% in into_heap.