lisp: Add simple expander, as an alternative to psyntax.
I noticed that very often, when I introduce a bug into the compiler or
VM, it sails past the one test that does not involve psyntax, and then
all other tests fail because psyntax triggers the bug. This makes simple
errors hard to debug, so in this commit we change many tests to run
without psyntax, and introduce a simplified expander to drive a few
others that only need simple (let), (cond), (and), and (or).