lisp: Add four builtins: read-line, try, parse, load.
`(read-line)` reads and returns a line of input from stdin.  A newline
character is included at the end of the line.  If there's no more input,
returns an empty string. (I think it can also just error out, if there's
something wrong with stdin.)

`(try thunk)` calls `thunk` without arguments, then returns the pair
`(ok . <value>)` if the thunk returns a value, and `(error . <message>)`
if the thunk fails. It would be good to support R6RS error handling
instead of having this builtin.

`(parse string)` tries to parse `string` as a sequence of Scheme <datum>
forms. On success, it returns the list `(ok form0 ...)`. On a parse error,
it returns a pair `(error . <message>)`. If `string` does not parse but
is the prefix of something that would parse, returns `(incomplete)`.

`(load filename env)` opens a file, reads the whole thing, and runs it
as Scheme code. The second argument is the environment in which to run
the code (like `eval`).